requirements.txt
# FairGuard Enterprise — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# ── Experiment Tracking ───────────────────────────────────────────────────────
mlflow>=2.11.0

# ── Google Cloud ──────────────────────────────────────────────────────────────
google-cloud-aiplatform>=1.48.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.20.0
google-auth>=2.28.0

# ── Pipeline Orchestration ────────────────────────────────────────────────────
kfp>=2.7.0
google-cloud-pipeline-components>=2.7.0

# ── Data Validation & Contracts ───────────────────────────────────────────────
pydantic>=2.6.0

# ── Anthropic SDK ─────────────────────────────────────────────────────────────
anthropic>=0.25.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0

# ── Notebook / WIT ───────────────────────────────────────────────────────────
notebook>=7.1.0
witwidget>=1.8.1

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0
black>=24.0.0
ruff>=0.3.0
requests
python-dotenv>=1.0.0

# ── AutoML & Gradient Boosters (V2 Upgrade) ───
optuna
xgboost
lightgbm
catboost